fix: 연결 가능한 역의 수를 체크하는 메소드의 반환 타입이 long은 너무 크다고 판단되어 int 로 변

+++ src/main/java/nextstep/subway/line/domain/Sections.java
@@ -72,16 +72,16 @@ public class Sections {
     }
 
     private void validateConnectableSection(Section section) {
-        long result = getConnectedStationCount(section);
+        int result = getConnectedStationCount(section);
 
         validateExistingSection(result);
         validateUnConnectableSection(result);
     }
 
-    private long getConnectedStationCount(Section section) {
+    private int getConnectedStationCount(Section section) {
         Stations stations = toStations();
 
-        return stations.get().stream()
+        return (int) stations.get().stream()
                 .filter(section::isIncludeStation)
                 .count();
     }
